tsconfig.json and others fixed
