+++ backend/functions/tsconfig.json
@@ -7,7 +7,9 @@
     "outDir": "lib",
     "sourceMap": true,
     "strict": true,
-    "target": "es2019"
+    "target": "es2019",
+    "resolveJsonModule": true,
+    "esModuleInterop": true
   },
   "compileOnSave": true,
   "include": ["src"]
